+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 - Added a Works For You container - sarah
 - Upgrade to RN 0.42-rc.3 - alloy
 - Made Gene view work with new relay-based infinite scroll grid - cab
+- Updated our fonts lib - orta
+- Converted to TypeScript - luc, sarah, alloy, orta
 
 ### 1.2.1